Refactor and enhance stock movement and article handling
- Added `ExternalDocument` property to `AddStockMovementCommand` and `StockMovement` classes.
- Modified endpoint in `AddStockMovementCommandHandler` to "articles/stocks/movements".
- Enhanced `GetDocumentPdfQuery` to support querying by document number and added URL query string generation.
- Updated `GetDocumentPdfQueryHandler` to use the new URL query method.
- Simplified byte conversion in `SerialPortService` and removed unnecessary casting.
- Refactored image processing methods in `Utils.cs` for readability.
- Removed redundant dynamic casting in `XmlToObjectParser`.
- Moved `_mediator` initialization in `ArticleField` and added necessary using directives.
- Simplified `QuantityIsValid` method in `ArticleField`.
- Added a new tab for unique articles in `ArticleModal`.
- Refactored error handling in `ArticlesPage` to use a centralized method.
- Simplified modal initialization across various pages.
- Enhanced `AddStockModal` to support selecting a document number from a file picker.
- Added buttons to `ArticleHistoryPage` and `StockMovementsPage` for PDF display.
- Refactored `PageNavigator` to use a factory method for button creation.
- Corrected namespace typos from `InputFieds` to `InputFields`.
- Centralized error handling in `Page` class.
- Added method to `FilePicker` to get file path for opening files.
- Adjusted layout of `PageNavigator` buttons for consistent spacing.
- Removed redundant dynamic casting in `XmlToObjectParser`.
- Added necessary using directives in various files.
- Removed `SelectedEntity` type casting in multiple files.
- Replaced `ShowApiErrorAlert` with `HandleErrorResult` in `PermissionsPage.cs`.
- Simplified `BorderWidth` property in `DocumentPaymentMethodsBox.cs`.
- Removed casting of predicate in `POSWindow.cs`.
- Introduced new classes and handlers for article serial number PDFs and stock movement updates.
- Created UI components for input fields and modals to update stock movements and unique articles.
- Added methods for validation and event handling to enhance user interaction.

+++ LogicPOS.UI/Application/Hardware/SerialPortService.cs
@@ -100,7 +100,7 @@ namespace LogicPOS.Utility
             for (int i = 0; i < msg.Length; i += 2)
                 //convert each set of 2 characters to a byte
                 //and add to the array
-                comBuffer[i / 2] = (byte)Convert.ToByte(msg.Substring(i, 2), 16);
+                comBuffer[i / 2] = Convert.ToByte(msg.Substring(i, 2), 16);
             //return the array
             return comBuffer;
         }
